requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
